fix(workspace): resolve clippy warnings in 9 more crates and fix workspace lint config
Fix clippy warnings across mockforge-core, mockforge-chaos,
mockforge-runtime-daemon, mockforge-tui, mockforge-analytics,
mockforge-pipelines, mockforge-federation, mockforge-tunnel,
and mockforge-scenarios.

Key fixes:
- Move 4 rustc lints from [workspace.lints.clippy] to [workspace.lints.rust]
  (dead_code, unused_variables, unused_imports, private_interfaces)
- Replace std::io::Error::new(ErrorKind::Other) with Error::other()
- Fix bool_assert_comparison, manual_clamp, derivable_impls patterns
- Convert &PathBuf params to &Path, fix only_used_in_recursion
- Add #[allow(dead_code)] for sqlx::FromRow structs with unread fields
- Remove unnecessary f64 cast, unused imports, and unused variables
- Add missing docs for public variants and struct fields

+++ crates/mockforge-core/src/ai_studio/org_controls_db.rs
@@ -464,7 +464,7 @@ impl OrgControlsAccessor for DbOrgControls {
         .bind(user_uuid)
         .bind(feature_name)
         .bind(tokens as i64)
-        .bind(cost_usd as f64)
+        .bind(cost_usd)
         .bind(metadata.unwrap_or_else(|| serde_json::json!({})))
         .execute(&self.pool)
         .await
@@ -507,6 +507,7 @@ impl OrgControlsAccessor for DbOrgControls {
 /// Database row for org_ai_budgets table
 #[cfg(feature = "database")]
 #[derive(sqlx::FromRow)]
+#[allow(dead_code)]
 struct BudgetRow {
     id: Uuid,
     org_id: Uuid,
@@ -522,6 +523,7 @@ struct BudgetRow {
 /// Database row for org_ai_rate_limits table
 #[cfg(feature = "database")]
 #[derive(sqlx::FromRow)]
+#[allow(dead_code)]
 struct RateLimitRow {
     id: Uuid,
     org_id: Uuid,
@@ -534,6 +536,7 @@ struct RateLimitRow {
 /// Database row for org_ai_feature_toggles table
 #[cfg(feature = "database")]
 #[derive(sqlx::FromRow)]
+#[allow(dead_code)]
 struct FeatureToggleRow {
     id: Uuid,
     org_id: Uuid,
